requirements.txt
absl-py==0.9.0
astor==0.8.1
boto==2.49.0
boto3==1.14.33
botocore==1.17.33
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
docutils==0.15.2
Flask==1.1.2
gast==0.2.2
gensim==3.8.3
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
nltk==3.5
numpy
opt-einsum==3.3.0
pandas==1.1.0
protobuf==3.12.4
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.24.0
s3transfer==0.3.3
scipy
six
smart-open
tensorboard==1.15.0
tensorflow-cpu==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
tqdm==4.48.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
